[Triton][Kernel] Refactor: organize triton rotary_embedding kernels into layers/triton_ops
Move Triton kernel rotary_embedding_triton() to layers/triton_ops/rotary_embedding.py
